+++ database/migrations/2025_05_31_013235_add_category_id_to_count_stock_table.php
@@ -11,12 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('count_stock', function (Blueprint $table) {
-            $table->engine = 'InnoDB';
-            $table->integer('category_id')->index('count_stock_category_id')->nullable()->after('warehouse_id');
-            $table->foreign('category_id', 'count_stock_category_id')->references('id')->on('categories')->onUpdate('RESTRICT')->onDelete('RESTRICT');
-
-        });
+        if (!Schema::hasColumn('count_stock', 'category_id')) {
+            Schema::table('count_stock', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->integer('category_id')->index('count_stock_category_id')->nullable()->after('warehouse_id');
+                $table->foreign('category_id', 'count_stock_category_id')->references('id')->on('categories')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            });
+        }
     }
 
     /**
